fix(객체로 저장 출력과 객체 분리):
김수호 그는 신입니다
