+++ lib/pages/cloth_input.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'dart:convert';
+import 'package:crush_client/pages/objects/cloth_object.dart';
 
 class ClothInput extends StatefulWidget {
   @override
@@ -178,31 +179,4 @@ class _ClothInputState extends State<ClothInput> {
     );
   }
 }
-class Cloth {
-  final String name;
-  final String color;
-  final String type;
-  final String thickness;
 
-  Cloth({
-    required this.name,
-    required this.color,
-    required this.type,
-    required this.thickness,
-  });
-
-  factory Cloth.fromJson(Map<String, dynamic> json) {
-    return Cloth(
-      name: json['name']??'No Name',
-      color: json['color'],
-      type: json['type'],
-      thickness: json['thickness'],
-    );
-  }
-  Map<String, dynamic> toJson() => {
-    'name': name,
-    'color': color,
-    'type': type,
-    'thickness': thickness,
-  };
-}
